configure Celery using factory design pattern
