+++ .env.example
@@ -1,0 +1,3 @@
+JWT_SECRET_KEY=''
+MONGO_PASSWORD=''
+FLASK_APP=''
