+++ pyproject.toml
@@ -17,6 +17,7 @@ pymongo = "^3.10.1"
 python-dotenv = "^0.10.3"
 beautifulsoup4 = "^4.8.2"
 requests = "^2.22.0"
+celery = "^4.4.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
