+++ recipebox/app.py
@@ -1,3 +1,3 @@
-from recipebox import create_app
+from recipebox.factory import create_app
 
 app = create_app('flask.cfg')
